Exposing GrpcResponse.grpcResponseToException
